gh-38: Protect auth-required controllers with decorators
